Python API: Moving handle structs from _k4a.py back to _k4atypes.py so that it aligns more with k4a.h and k4atypes.h.
